Adjustments 3 for shorten-64-to-32 warnings, issue #710

+++ include/seqan/system/file_async.h
@@ -1047,7 +1047,7 @@ namespace SEQAN_NAMESPACE_MAIN
         return aio_error(&request);
     }
 
-    inline int _returnValue(AiocbWrapper & request) {
+    inline ssize_t _returnValue(AiocbWrapper & request) {
 //IOREV _nodoc_
         return aio_return(&request);
     }
